Separate MurmurHash32 from hash_functions.cuh
