+++ cpp/src/hash/concurrent_unordered_map.cuh
@@ -21,7 +21,7 @@
 #include <hash/managed.cuh>
 
 #include <cudf/detail/nvtx/ranges.hpp>
-#include <cudf/detail/utilities/hash_functions.cuh>
+#include <cudf/hashing/detail/murmur_hash32.cuh>
 #include <cudf/utilities/default_stream.hpp>
 #include <cudf/utilities/error.hpp>
