+++ cpp/src/io/json/json_tree.cu
@@ -19,12 +19,12 @@
 #include <hash/helper_functions.cuh>
 #include <io/utilities/hostdevice_vector.hpp>
 
-#include <cudf/detail/hashing.hpp>
 #include <cudf/detail/nvtx/ranges.hpp>
 #include <cudf/detail/scatter.cuh>
 #include <cudf/detail/utilities/algorithm.cuh>
-#include <cudf/detail/utilities/hash_functions.cuh>
 #include <cudf/detail/utilities/vector_factories.hpp>
+#include <cudf/hashing/detail/hashing.hpp>
+#include <cudf/hashing/detail/murmur_hash32.cuh>
 #include <cudf/utilities/error.hpp>
 #include <cudf/utilities/span.hpp>
