+++ cpp/src/stream_compaction/stream_compaction_common.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include <cudf/detail/utilities/hash_functions.cuh>
+#include <cudf/hashing/detail/hash_functions.cuh>
 #include <cudf/table/row_operators.cuh>
 #include <cudf/table/table_device_view.cuh>
